feat(mobile): add EntryNavigationHeader and refactor EntryTitle
- Introduced EntryNavigationHeader component for improved navigation and header management in entry detail screens.
- Added titleHeightAtom to EntryContentContext for dynamic title height management.
- Refactored EntryTitle to utilize the new context and simplify layout handling.
- Updated EntryDetailScreen to integrate EntryNavigationHeader and enhance user experience.
