+++ apps/mobile/src/modules/entry-content/ctx.ts
@@ -5,6 +5,7 @@ interface EntryContentContextType {
   showAISummaryAtom: PrimitiveAtom<boolean>
   showAITranslationAtom: PrimitiveAtom<boolean>
   showReadabilityAtom: PrimitiveAtom<boolean>
+  titleHeightAtom: PrimitiveAtom<number>
 }
 export const EntryContentContext = createContext<EntryContentContextType>(null!)
 export const useEntryContentContext = () => {
